Revert "add tracing integration test (#156)"
This reverts commit 7c8fc1c5642c6fca8aa97697f3895d89ef4c6956.
