+++ tracing/dubbo/go-client/app/user.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package pkg
+package main
 
 import (
 	"context"
@@ -23,18 +23,16 @@ import (
 )
 
 import (
-	"github.com/dubbogo/gost/log"
-	"github.com/opentracing/opentracing-go"
-)
-
-import (
 	"dubbo.apache.org/dubbo-go/v3/config"
 	hessian "github.com/apache/dubbo-go-hessian2"
 )
 
+var (
+	userProvider = &UserProvider{}
+)
+
 func init() {
-	config.SetProviderService(new(UserProvider))
-	// ------for hessian2------
+	config.SetConsumerService(userProvider)
 	hessian.RegisterPOJO(&User{})
 }
 
@@ -46,22 +44,13 @@ type User struct {
 }
 
 type UserProvider struct {
-}
-
-func (u *UserProvider) GetUser(ctx context.Context, req []interface{}) (*User, error) {
-	gxlog.CInfo("req:%#v", req)
-	rsp := User{"A001", "Alex Stocks", 18, time.Now()}
-	span, ctx := opentracing.StartSpanFromContext(ctx, "User-Server-Span")
-	time.Sleep(100 * time.Millisecond)
-	span.Finish()
-	gxlog.CInfo("rsp:%#v", rsp)
-	return &rsp, nil
+	GetUser func(ctx context.Context, req []interface{}, rsp *User) error
 }
 
 func (u *UserProvider) Reference() string {
 	return "UserProvider"
 }
 
-func (u User) JavaClassName() string {
+func (User) JavaClassName() string {
 	return "com.ikurento.user.User"
 }
