+++ tracing/dubbo/go-server/app/user.go
@@ -15,12 +15,28 @@
  * limitations under the License.
  */
 
-package pkg
+package main
 
 import (
 	"context"
 	"time"
 )
+
+import (
+	"github.com/dubbogo/gost/log"
+	"github.com/opentracing/opentracing-go"
+)
+
+import (
+	"dubbo.apache.org/dubbo-go/v3/config"
+	hessian "github.com/apache/dubbo-go-hessian2"
+)
+
+func init() {
+	config.SetProviderService(new(UserProvider))
+	// ------for hessian2------
+	hessian.RegisterPOJO(&User{})
+}
 
 type User struct {
 	ID   string
@@ -30,13 +46,22 @@ type User struct {
 }
 
 type UserProvider struct {
-	GetUser func(ctx context.Context, req []interface{}, rsp *User) error
+}
+
+func (u *UserProvider) GetUser(ctx context.Context, req []interface{}) (*User, error) {
+	gxlog.CInfo("req:%#v", req)
+	rsp := User{"A001", "Alex Stocks", 18, time.Now()}
+	span, ctx := opentracing.StartSpanFromContext(ctx, "User-Server-Span")
+	time.Sleep(100 * time.Millisecond)
+	span.Finish()
+	gxlog.CInfo("rsp:%#v", rsp)
+	return &rsp, nil
 }
 
 func (u *UserProvider) Reference() string {
 	return "UserProvider"
 }
 
-func (User) JavaClassName() string {
+func (u User) JavaClassName() string {
 	return "com.ikurento.user.User"
 }
