--- tracing/dubbo/go-server/docker/docker-compose.yml
@@ -1,8 +1,0 @@
-version: '3'
-
-services:
-  zookeeper:
-    image: zookeeper
-    ports:
-      - 2181:2181
-    restart: on-failure
