--- tracing/dubbo/go-server/tests/integration/userprovider_test.go
@@ -1,74 +1,0 @@
-// +build integration
-
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package integration
-
-import (
-	"context"
-	"testing"
-)
-import (
-	"github.com/dubbogo/gost/log"
-	"github.com/opentracing/opentracing-go"
-	zipkinot "github.com/openzipkin-contrib/zipkin-go-opentracing"
-	"github.com/openzipkin/zipkin-go"
-	zipkinhttp "github.com/openzipkin/zipkin-go/reporter/http"
-	"github.com/stretchr/testify/assert"
-
-	"dubbo.apache.org/dubbo-go/v3/common/logger"
-)
-
-func TestGetUser(t *testing.T) {
-	initZipkin()
-	gxlog.CInfo("\n\n\nstart to test dubbo")
-	user := &User{}
-	span, ctx := opentracing.StartSpanFromContext(context.Background(), "Test-Client-Service")
-	err := userProvider.GetUser(ctx, []interface{}{"A001"}, user)
-	span.Finish()
-	assert.Nil(t, err)
-
-	gxlog.CInfo("response result: %v\n", user)
-	//assert.Equal(t, "A001", user.Id)
-	assert.Equal(t, "Alex Stocks", user.Name)
-	assert.Equal(t, int32(18), user.Age)
-	assert.NotNil(t, user.Time)
-}
-
-func initZipkin() {
-	// set up a span reporter
-	reporter := zipkinhttp.NewReporter("http://localhost:9411/api/v2/spans")
-
-	// create our local service endpoint
-	endpoint, err := zipkin.NewEndpoint("myService", "myservice.mydomain.com:80")
-	if err != nil {
-		logger.Errorf("unable to create local endpoint: %+v\n", err)
-	}
-
-	// initialize our tracer
-	nativeTracer, err := zipkin.NewTracer(reporter, zipkin.WithLocalEndpoint(endpoint))
-	if err != nil {
-		logger.Errorf("unable to create tracer: %+v\n", err)
-	}
-
-	// use zipkin-go-opentracing to wrap our tracer
-	tracer := zipkinot.Wrap(nativeTracer)
-
-	// optionally set as Global OpenTracing tracer instance
-	opentracing.SetGlobalTracer(tracer)
-}
